Rename 'hand_open 3:4.svg' to 'hand_open_3_4.svg'. (#10)
Oops, @ninapaley you added a file with a colon in it, which is bad for Windows systems because that's a special character that can't appear in filenames, so it broke my Windows setup. Fortunately I was able to rename the file on a Linux system but to prevent weird stuff like this from happening in the future, it'd probably be best if you could limit the names of your SVG files to something that works across all kinds of operating systems--maybe the roman alphabet, numbers, hyphens and/or dashes?