Add animated GIF export. (#79)
This adds an "Export GIF" button _if and only if_ animation is enabled on the Mandala page.

The GIF has the same dimensions as the canvas at the time that the button was clicked.

Currently the animation is locked to 15 FPS, which I think might be the maximum rate that some platforms limit it to, but I'm not sure.  We can always adjust it in the future.